+++ src/components/schedule/schedule-edit-panel.tsx
@@ -8,7 +8,7 @@ import {
   useUpdateSchedule,
 } from "@/hooks/schedule";
 import { Button, Textarea } from "@/components/shadcn";
-import ScheduleAddInputFields from "./schedule-add-input-fields";
+import ScheduleAddInputFields from "./schedule-input-fields";
 import { formatSelectedDateTime } from "@/utils/time/dateTime";
 
 import { useEffect } from "react";
