+++ src/hooks/schedule/index.ts
@@ -9,3 +9,4 @@ export * from "./use-create-schedule";
 export * from "./use-schedule-list";
 export * from "./use-today-tomorrow-schedules";
 export * from "./use-update-schedule";
+export * from "./use-daum-post-code";
