+++ src/hooks/shared/use-enter-to-focus-next.ts
@@ -1,0 +1,15 @@
+import { useRef } from "react";
+
+export function useEnterToFocusNext() {
+  const inputRefs = useRef<(HTMLInputElement | null)[]>([]);
+
+  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>, index: number) => {
+    if (e.key === "Enter") {
+      e.preventDefault();
+      const next = inputRefs.current[index + 1];
+      if (next) next.focus();
+    }
+  };
+
+  return { inputRefs, handleKeyDown };
+}
